requirements.txt
streamlit==1.17.0
pandas==1.5.3
numpy==1.24.1
plotly==5.13.0
matplotlib==3.6.3
pillow==9.4.0